Use cabin 0.7.2 instead of 0.8.0
This appears to fix issue #362.

The symptom: Cabin-related stacktrace during fpm invocation.

Cabin 0.8.0 changes its API such that it is incompatible with the
version of fpm in use in chef-bach.  Downgrading to cabin 0.7.2
resolves the API incompatibility.

Since Cabin is a log library, I expected to find an underlying
problem, but none presented itself.